requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
streamlit
Pillow
pytesseract
pdfminer.six
pypdfium2
requests
numpy
fpdf
pypdf 
faiss-cpu
langchain
euriai
langchain_community
sentence_transformers
